requirements.txt
streamlit
langchain
langchain-community
google-generativeai
faiss-cpu
python-dotenv
